requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
bcrypt==4.2.0
beanie==1.27.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
databases==0.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.3
fastapi-cache2==0.2.2
flake8==7.1.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
lazy-model==0.2.0
loguru==0.7.3
makefun==1.15.6
Mako==1.3.6
MarkupSafe==3.0.2
mccabe==0.7.0
motor==3.6.0
packaging==24.2
passlib==1.7.4
pendulum==3.0.0
pluggy==1.5.0
psycopg2==2.9.10
pwdlib==0.2.1
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
pyflakes==3.2.0
PyJWT==2.9.0
pymongo==4.9.2
pytest==8.3.4
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
redis==4.6.0
requests==2.32.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
time-machine==2.16.0
toml==0.10.2
transliterate==1.10.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
